+++ solvers/s0002.c
@@ -23,5 +23,5 @@ bool solve0002(char *buffer, size_t buffer_size)
         current = new;
     }
     int ret = snprintf(buffer, buffer_size, "%d", sum);
-    return ret >= 0 && ret < buffer_size;
+    return ret >= 0 && (unsigned int)ret < buffer_size;
 }
