+++ solvers/s0004.c
@@ -35,5 +35,5 @@ bool solve0004(char *buffer, size_t buffer_size)
         }
     }
     int ret = snprintf(buffer, buffer_size, "%d", largest_palindromic_number);
-    return ret >= 0 && ret < buffer_size;
+    return ret >= 0 && (unsigned int)ret < buffer_size;
 }
